requirements.txt
streamlit==1.41.1
google-generativeai==0.8.3
python-dotenv==1.0.1
pypdf==5.1.0
langchain_community==0.3.14
pinecone==5.4.2